refactor: extract brush, zoom, and range-slider from hooks to core
