+++ src/core/brush/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
 export * from './utils';
-export * from './zoom';
