+++ src/core/brush/types.ts
@@ -1,8 +1,7 @@
 import type {BrushBehavior} from 'd3-brush';
 
-import type {ZoomType} from '~core/constants';
-
-import type {ChartBrush, DeepRequired} from '../../types';
+import type {ZoomType} from '../constants';
+import type {ChartBrush, DeepRequired} from '../types';
 
 export type BrushSelection = [number, number] | [[number, number], [number, number]];
