+++ src/core/range-slider/types.ts
@@ -1,0 +1,4 @@
+export type RangeSliderState = {
+    max: number;
+    min: number;
+};
