+++ src/hooks/index.ts
@@ -3,7 +3,7 @@ export * from '~core/axes/types';
 export * from './useAxisScales';
 export * from '~core/scales/types';
 export * from './useBrush';
-export * from './useBrush/types';
+export * from '~core/brush/types';
 export * from './useCrosshair';
 export * from './usePrevious';
 export * from './useRangeSlider';
@@ -15,5 +15,5 @@ export * from '~core/layout/split';
 export * from '~core/layout/split-types';
 export * from './useTooltip';
 export * from './useZoom';
-export * from './useZoom/types';
+export * from '~core/zoom/types';
 export * from './types';
