+++ src/hooks/types.ts
@@ -1,22 +1,8 @@
 import type {TextRowData} from '../components/types';
-import type {
-    BaseTextStyle,
-    ChartBrush,
-    ChartData,
-    ChartMargin,
-    ChartTitle,
-    ChartZoom,
-    DeepRequired,
-} from '../types';
+import type {BaseTextStyle, ChartData, ChartTitle} from '../types';
 import type {HtmlItem} from '../types/chart-ui';
 
-export type PreparedZoom = DeepRequired<Omit<ChartZoom, 'enabled' | 'brush'>> &
-    DeepRequired<{brush: ChartBrush}>;
-
-export type PreparedChart = {
-    margin: ChartMargin;
-    zoom: PreparedZoom | null;
-};
+export type {PreparedChart, PreparedZoom} from '~core/chart/types';
 
 export type PreparedTitle = Omit<ChartTitle, 'margin' | 'style'> & {
     height: number;
