+++ src/hooks/useRangeSlider/types.ts
@@ -1,15 +1,13 @@
+import type {PreparedRangeSlider, PreparedXAxis, PreparedYAxis} from '~core/axes/types';
+import type {BrushSelection, UseBrushProps} from '~core/brush/types';
+import type {PreparedChart} from '~core/chart/types';
+import type {RangeSliderState} from '~core/range-slider/types';
 import type {ChartScale} from '~core/scales/types';
 import type {PreparedLegend, PreparedSeries, PreparedSeriesOptions} from '~core/series/types';
 
 import type {ChartXAxis, ChartYAxis, LegendConfig} from '../../types';
-import type {PreparedChart} from '../types';
-import type {PreparedRangeSlider, PreparedXAxis, PreparedYAxis} from '../useAxis/types';
-import type {BrushSelection, UseBrushProps} from '../useBrush/types';
 
-export type RangeSliderState = {
-    max: number;
-    min: number;
-};
+export type {RangeSliderState};
 
 export interface RangeSliderProps {
     activeLegendItems: string[];
